Backend/requirements.txt
Flask_SocketIO==4.3.1
Flask==1.1.2
Flask-Session==0.3.2
Jinja2==3.0.3
itsdangerous==2.0.1
werkzeug==2.0.3
python-engineio==3.13.2
python-socketio==4.6.0
flask_restful==0.3.9
firebase_admin==5.2.0
python-dotenv==0.19.2
